crisis-bot/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
websockets==14.1
twilio==9.4.0
google-cloud-firestore==2.19.0
python-dotenv==1.0.1
openai==1.59.7
azure-cosmos==4.9.0
azure-servicebus==7.13.0
azure-identity==1.19.0
azure-cognitiveservices-speech==1.42.0
audioop-lts; python_version >= "3.13"
